Moved IIsPunchItemCommand to Equinor.ProCoSys.Completion.Command.PunchItemCommands ns
